feat (calypte): refactoring renaming
