--- backend/apps/calypte_api/calypte_api/common/__init__.py
@@ -1,6 +1,0 @@
-from calypte_api.common.models import device_tag_lookup
-
-
-__all__ = [
-    "device_tag_lookup",
-]
