+++ backend/apps/calypte_api/calypte_api/common/exceptions.py
